requirements.txt
# Análise de Extratos Bancários - Dependências

# Leitura de PDFs
PyPDF2==3.0.1
pdfplumber==0.11.4

# Análise de dados
pandas==2.2.3
numpy==2.1.3

# Processamento de texto
python-dateutil==2.9.0

# CLI e interface
click==8.1.7
rich==13.9.4

# Validação de dados
pydantic

# Testes
pytest==8.3.4
pytest-cov==6.0.0

# Desenvolvimento
black==24.10.0
flake8==7.1.1
mypy==1.13.0

# Utilitários
python-dotenv==1.0.1

#RAG MCP
mcp
abacusai
watchdog
faiss-cpu

# Embeddings e modelos
sentence-transformers==3.0.1
torch==2.4.0